Take out rpc.py
